refactor: introduce chat completion provider abstraction
