+++ Services/Interfaces/IAnthropicProvider.cs
@@ -1,0 +1,10 @@
+namespace DeepSeek_v4_for_VisualStudio.Services
+{
+    /// <summary>
+    /// Anthropic Messages API Provider 插槽。
+    /// 当前仅保留协议边界，具体请求/响应映射与工具调用适配后续实现。
+    /// </summary>
+    public interface IAnthropicProvider : IChatCompletionProvider
+    {
+    }
+}
