+++ Services/Interfaces/IResponsesApiProvider.cs
@@ -1,0 +1,10 @@
+namespace DeepSeek_v4_for_VisualStudio.Services
+{
+    /// <summary>
+    /// OpenAI Responses API Provider 插槽。
+    /// 当前仅保留协议边界，具体事件流、工具调用和状态映射后续实现。
+    /// </summary>
+    public interface IResponsesApiProvider : IChatCompletionProvider
+    {
+    }
+}
